Added option to use absolute time
Whenthe option absoluteTime is used, temporal paths will be detected
only if integer time steps are consecutive, like in the two consecutive
time stamped edges (a,b,1) and (b,c,2). The sequence (a,b,1) and (b.c,3)
will not be detected as a path. If this option is switched off, the
latter case will be detected as a temporal path if no event happened
between 1 and 3, i.e. time is only used to construct a logical order.